Audio System & Amplitude Audio plugins (#1)
* First ever implementation of an AudioSystem

This commit introduces the first version of an AudioSystem. The AudioSystem is a plugin abstracting most of the features provided by audio engines like Wwise, FMOD or Amplitude. But actually, any sound engine can be used as middleware, even really small ones, like SoLoud.

This commit adds a working version of the AudioSystem with support of triggering audio events from the middleware.

The AudioSystem is splitted in a runtime plugin (ezAudioSystemPlugin), an editor plugin (ezEditorPluginAudioSystem), and a static library (ezSharedPluginAudioSystem) which will be used by the middleware editor plugin to implement required editor actions (eg: reload controls).

* Add more options to ezAudioControlsComponent

- Added an `AutoLoad` property to define if the resource should be loaded with the component initialization.
- Added a scriptable `Load` function to manually load the controls in the Audio System.
- Added a scriptable `Unload` function to manually unload controls.
- Ensure the controls are unloaded with the component deinitalization.

* Implement RTPC

Real-Time Parameter Controls are used by the game engine to adjust some values in the audio engine. This commit implements the support of RTPC in the Audio System and exposes the functionality to the user through a `ezAudioRtpcComponent`.

* Simplified code using macros

* Implement Audio Listeners

The support of audio listeners has been added. The Audio System will register a default editor listener when compiling for development.

* Apply offset to listener position.

* Optimized game object to audio entity references

Now audio entities are managed through a single game object with a single audio proxy component. All audio components in the same component will then share the same entity using the correct audio proxy.

* Implement an Audio Animation Component

The Audio Animation Component listens to animation events triggered by the same game object to activate a specific trigger. It can also use a specific joint as audio entity so the activated trigger will follow the assigned bone transformation according to the animation.

* Ensure that the audio system is initialized.

* Make the audio system final.

* Be less verbose.

* Optimize audio system startup and shutdown.

Now it's the responsibility of the audio middleware plugin to startup the audio system.

* Specialize update function names.

* Audio Switch State Component

- Implemented the Audio Switch State component
- Optimized the audio controls asset to clearly separate different types of controls.
- Fixed issue on audio proxy dependent components being hidden in the editor.

* Implement Audio environments

Environments are used to apply environmental effects when a sound is in a specific area in the game.

* Fix playback issues on Audio Trigger

* Fix Release build

* Optimize the way to get environment ID from component.

* Add missing static link references

* Implement Unload Trigger request.

* Implement AudioBoxEnvironment

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>

* Refactor Audio System.

Audio control assets are now not built by the audio system itself. The plugin will give the audio middlewares the responsibility to feed the audio system with audio controls the way they want, without the explicit need of audio controls.

* Implement editor actions

Now the user is now able to mute, unmute, and change the volume of the audio system through the editor.

* Compute Obstruction and Occlusion values for each audio trigger.

Also fixed some issues with the request callback, by making sure synchronous requests are executed with their callback before running asynchronous requests.

* Initial implementation of Amplitude Audio as a plugin

The plugin uses the Audio System as a base to communicate with EZ.

* Some cleanup.

* Remove reference to unused plugin.

* Cleanup.

* Unregister controls on collection asset unload.

* Add ability to get the master gain and the muted state from the middleware.

* Update to the new API.

* Enable sound bank registration in the audio system.

Also moved specific Amplitude code from the audio system plugin to the Amplitude plugin.

* Load and unload banks.

* Use ezRefCounted on proxy components.

* Fix allocation issues when the same trigger is loaded many times with the same event.

* Add registration of sound banks into audio controls collection assets.

* Fix missing include.

* Add translations.

* Add component icons.

* Don't cast occlusion rays if there is no direction.

* Load the audio system configuration file from the audio system itself.

* Correctly set and unset default listener.

* Ensure audio proxies are properly unregistered.

Fixed an issue where audio proxies were not properly unregistered once unreferenced by all dependents. Also send all audio requests generated by audio proxies at once in the update method.

* Set up editor actions for Amplitude Audio plugin.

A button to reload Amplitude controls from the editor has been added in the scene toolbar and in the audio controls collection asset document toolbar as well.

* Remove unused fields.

* Add the Amplitude settings dialog.

- Allow to customize the Amplitude settings with a settings dialog.
- Add more translations.
- Add Amplitude icons for components and assets.
- Remove the capacity to save middleware settings via engine plugins, middleware implementations should provide a way to save settings through editor plugins or whatever.

* Add scene editor icons for audio system components.

* Set a specific exclusive feature for audio system middlewares.

This allows to not select more than one audio system middleware in the plugins selection dialog.

* Add sound banks during controls collection transform process.

* Optimize PNG icons.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>