Implement Upscale Pass.
This is using the AMD Fidelity FX Super Resolution library. to offer up to 2x upscaling with minimal costs. We should figure out a way to detect the user graphic card to switch between FP16 and FP32 precision, but for now, FP32 precision has been hard-coded since it's widely supported (but slower than FP16). A simple Bilinear upscaler is also implemented in cases FSR is disabled.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>