wip(RPI): RPI architecture improvements.
The RenderSystem is now limited to only register SceneContext to a World.

The SceneContext will now reference a single World, so we can have many SceneContext registered to the RenderSystem according to each World in need of rendering features.

The Collect, Extract, and Prepare phases has been moved to the SceneContext, since each of them may process different data per World.

The place where to collect RenderObject per views and stages is still not yet decided.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>