wip(RPI): New shaders system.
The new shaders system is based on the Slang shading language, which provides much more features, link-time specializations, and supports all the RHI platforms. I will sadly drop the SPSL support as besides the fact it provides nice features, it is not that mature compared to Slang, and I start to not have enough time to invest in it.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>