wip(RPI): Add dynamic render object instantiation.
The RPI can now dynamically generate instances (for instanced rendering) from render data. This is done by the render stage before render operations. The render stage pass through final render objects and batch them in instances if needed, and then only the batched objects are rendered with instanced rendering.

It's to the render object implementation to hold instance data (via the `Instantiate()` method), and to its associated render feature to properly handle it.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>