feat(RHIMTL): Improved CPU load by using argument buffers.
Argument buffers allows to bind a collection of resources once to a single buffer. This reduces a lot the CPU. The concept of argument buffers in Metal is similar to descriptor sets for Vulkan or descriptor tables for D3D12. There is no counterpart for D3D11 since the concepts of `sets` has been introduced to D3D12.

Signed-off-by: Axel Nana <axel.nana@aliens-group.com>